requirements.txt
yfinance
pandas
pandas-ta
gspread
oauth2client
line-bot-sdk
python-dotenv
